Renamed ppork3 library path to simply lib
